+++ src/components/tables/reports/revenue.tsx
@@ -1,20 +1,21 @@
 import { useCallback, useState } from "react";
 import Filter from "../../filterAndSort/filter";
 // import Search from "../../inputs/search";
-import { apartmentById } from "../../../types/apiData/apartment";
 import LoadingButton from "../../button";
-import ExportSelect from "../../inputs/select/exportSelect";
+// import { apartmentById } from "../../../types/apiData/apartment";
+// import ExportSelect from "../../inputs/select/exportSelect";
 import { revenueReportList } from "../../../types/apiData/reports";
 import NoResult from "../../noResult";
 import Pagination from "../../pagination";
 import formatDate from "../../../utils/isoDateConverter";
 import useGetRevenueReport from "../../../services-hooks/reports/revenue";
 import useGetReportSummary from "../../../services-hooks/reports/report-summary";
-import ApartmentSingleSearch from "../../inputs/search/apartment-single-search";
+// import ApartmentSingleSearch from "../../inputs/search/apartment-single-search";
 import { useAppDispatch } from "../../../stores/hooks";
 import { openSnackbar } from "../../../stores/appFunctionality/snackbar";
 import ExportToCSV from "../../export-to-csv";
 import { revenueReportExportFormater } from "../../../utils/export-formerter-functions";
+import ApartmentThroughBuildingSelector from "../../inputs/select/apartment-through-building-selector";
 
 export default function RevenueReportTable() {
   const [filterDates, setFilterDates] = useState<{
@@ -22,32 +23,25 @@ export default function RevenueReportTable() {
     end_date: string;
   }>();
   const dispatch = useAppDispatch();
-  const [apartment, setApartment] = useState<apartmentById>({} as any);
+  // const [apartment, setApartment] = useState<apartmentById>({} as any);
   const [currentPage, setCurrentPage] = useState(1);
 
-  const { data, isLoading, isFailed, setIsFailed, retryFunction, pagination } =
-    useGetRevenueReport({
-      page: currentPage,
-      apartmentId: String(apartment?.id || ""),
-      start_date: filterDates?.start_date,
-      end_date: filterDates?.end_date,
-    });
+  const [apartmentId, setApartmentId] = useState("");
+  const [buildingId, setBuildingId] = useState("");
+
+  const { data, isLoading, retryFunction, pagination } = useGetRevenueReport({
+    page: currentPage,
+    apartmentId: buildingId && apartmentId ? String(apartmentId) : "",
+    start_date: filterDates?.start_date,
+    end_date: filterDates?.end_date,
+  });
 
   const handleLoadData = useCallback(() => {
-    if (apartment?.id) {
-      retryFunction();
-    } else {
-      dispatch(
-        openSnackbar({
-          message: "Please select an apartment to view its reports",
-          isError: true,
-        })
-      );
-    }
-  }, [apartment?.id]);
+    retryFunction();
+  }, [apartmentId]);
 
   const { data: reportSummary } = useGetReportSummary({
-    apartmentId: String(apartment?.id || ""),
+    apartmentId: String(apartmentId || ""),
     start_date: filterDates?.start_date,
     end_date: filterDates?.end_date,
   });
@@ -64,11 +58,15 @@ export default function RevenueReportTable() {
         <div>
           <Filter actionHandler={handleCustomersFiltering} />
         </div>
-        <ApartmentSingleSearch
-          placeholder="Apartment name..."
-          selected={apartment}
-          setSelected={setApartment}
-        />
+        <div className="w-full max-w-screen-md flex items-center flex-col md:flex-row gap-2">
+          <ApartmentThroughBuildingSelector
+            setApartmentId={setApartmentId}
+            apartmentId={apartmentId}
+            buildingId={buildingId}
+            setBuildingId={setBuildingId}
+            withLabel={false}
+          />
+        </div>
         <div className=" flex items-center gap-4">
           <LoadingButton
             label="Load data"