+++ src/pages/bookings/all-bookings/index.tsx
@@ -6,7 +6,6 @@ import CalendarIcon from "../../../assets/icons/calendar";
 import LinkButton from "../../../components/button/linkButton";
 import PlusIcon from "../../../assets/icons/plus";
 import AllBookingsListTable from "../../../components/tables/allBookingsLists";
-import ExportSelect from "../../../components/inputs/select/exportSelect";
 import useGetResourceAccessChecker from "../../../utils/admin/useAccessChecker";
 
 const breadCrumb = [